Don't risk missing events in gather_job_info
Seen with syndic, but in theory affects all setups. When a we call
gather_job_info if an event comes in for the main job we can discard the
event whilst we are looking for the gather_job_info event. To work
around I've made gather_job_info add events for the main job to the
pending_events list. get_event now always checks that list for events
matching it's tag