Remove deepcopy from merge_recurse
Try and find all the places where it's beneficial to deepcopy the first
input and do so.

In the pillar generation I couldn't tell what was going on with
opts['pillar'] I gave up and just deep copied it whenever it was accessed
I think this is probably more than was necessary.

In the slsutil module I've been conservative and deep copied the second
input given that we merge this into an empty dict (so there's no point
deep copying the first input) then merge further dicts so it would be
possible to modify a earlier dict in the list with a later one if they
shared some keys. I believe this is still more efficient than the
previous code, we have the same number of calls to deepcopy but the dict
being copied should be smaller in each case.